Simplify knowing if landmarks will arrive
Previously the simulator would only send a landmarks message if at least
one landmark was visible. This meant that if you didn't receive a
landmarks message, you couldn't be sure whether it was just delayed in
arriving or whether it just wasn't going to be sent. Now we just send an
empty landmarks message if there is no landmark visible.